packages/python/pyproject.toml
[project]
name = "model-metadata-central"
version = "0.1.0"
description = "Typed registry of LLM model metadata for Python — context windows, pricing, provider routing."
readme = "README.md"
license = "MIT"
authors = [
    { name = "Eric Allen", email = "era@lakera.ai" }
]
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.0,<3",
    "pyyaml>=6.0,<7",
]
keywords = ["llm", "language model", "metadata", "openai", "anthropic", "pricing", "context-window"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries",
]

[project.urls]
Homepage = "https://github.com/InterwebAlchemy/model-metadata-central"
Repository = "https://github.com/InterwebAlchemy/model-metadata-central"
Issues = "https://github.com/InterwebAlchemy/model-metadata-central/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["model_metadata_central"]

[dependency-groups]
dev = [
    "black>=23.11.0",
    "isort>=5.12.0",
    "jsonschema>=4.19.2",
    "pre-commit>=3.5.0",
    "pytest>=7.4.3",
]

[tool.black]
target-version = ["py311"]

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
